Split kanayago.patch file to more easier maintain multi Ruby Parser
patch/3.4/kanayago.patch for Ruby 3.4 Parser.
patch/head/kanayago.patch for Ruby head Parser.

Only one patch file too difficult to maintain.
And, If follow the parser changes in Ruby head, the differences may differ from the changes to the parser in Ruby 3.4.

Also, since there is a possibility that the parser changes to Ruby 3.4 will be backported.
Then, need to make it easier to support those backports.
So, separated the patches for each version.